+++ cyberdrop_dl/config/__init__.py
@@ -9,7 +9,8 @@ from cyclopts.bind import normalize_tokens
 from pydantic import AfterValidator, BaseModel, Field, NonNegativeInt, PositiveInt
 
 from cyberdrop_dl import yaml
-from cyberdrop_dl.constants import DEFAULT_DOWNLOAD_STORAGE
+from cyberdrop_dl.config.appdata import AppData
+from cyberdrop_dl.constants import DEFAULT_DOWNLOAD_PATH
 from cyberdrop_dl.exceptions import CDLConfigRuntimeErrorsGroup
 from cyberdrop_dl.models import DeferedModel
 from cyberdrop_dl.models.types import ByteSizeSerilized, FalsyAsTuple  # noqa: TC001
@@ -56,7 +57,7 @@ class Config(DeferedModel, title="cyberdrop-dl config"):
     deep_scrape: bool = False
     delete_empty_folders: bool = True
     delete_partial_files: bool = False
-    download_folder: Annotated[Path, Parameter(alias=("--output", "-o", "-d"))] = DEFAULT_DOWNLOAD_STORAGE
+    download_folder: Annotated[Path, Parameter(alias=("--output", "-o", "-d"))] = DEFAULT_DOWNLOAD_PATH
     downloads: Downloads = Field(default_factory=Downloads)
     dump_json: Annotated[bool, Parameter(alias="-j")] = False
     filters: Filters = Field(default_factory=Filters)
@@ -120,11 +121,12 @@ class Config(DeferedModel, title="cyberdrop-dl config"):
         if self._resolved:
             return
 
-        self.logs.resolve_filenames()
+        default_log_folder = AppData.default().logs_folder
+        self.logs.resolve_filenames(default_log_folder)
         _resolve_paths(self)
         if self.logs.expire_after:
             self.logs.delete_old_logs_and_folders()
-            cleanup.rm_empty_dirs(self.logs.folder)
+            cleanup.rm_empty_dirs(self.logs.effective_log_folder)
         self._resolved = True