+++ tests/conftest.py
@@ -5,6 +5,7 @@ from typing import TYPE_CHECKING
 
 import pytest
 
+from cyberdrop_dl.config.appdata import AppData, AppDirs
 from cyberdrop_dl.manager import Manager
 
 if TYPE_CHECKING:
@@ -61,8 +62,12 @@ async def logs(caplog: pytest.LogCaptureFixture) -> pytest.LogCaptureFixture:
 
 
 @pytest.fixture
-def manager() -> Generator[Manager]:
-    with Manager()() as manager:
+def manager(tmp_cwd: Path) -> Generator[Manager]:
+    with Manager(
+        appdata=AppData.from_dirs(
+            AppDirs.from_path(tmp_cwd / "pytest_appdata"),
+        )
+    )() as manager:
         yield manager